Operators: Added Operator Chains and String Support
Currently operators cannot be used inside of function calls and they
don't handle parenteses. Additionally they only search when predefined
tokens are found (operator tokens). This means that operators have to
use one of the already known tokens. This is going to change but I'm not
sure when.

Also, Added String Literals