Operators: Kinda Fixed Issue With Multiple Expressions In Operator
The issue was with two expression in an operator in a row without a
landmark. e.g.:
oper <int a int b +>
1 2 +
Would not find the operator call because it would try to get "1 2" as a
single expression. The pseudo fix I made is to check how many expression
we should find and if it matches just give them directly to the operator
call. This doesn't fix the issue when we have mutliple operators in the
call e.g.:
1 2 + 1 +
This is kind of weird and I don't see it being much of an issue but I'd
still like it to work. I think the right fix for this is to detect when
we could have a combination like 1 + 1 -> 2 and then search for the
pattern after but I'm not sure ATM.