Parser: Fixed Parsing Complex Function Calls
Previously operator use within function calls.
e.g. print(1 + 2) or print(print(1))
This works now. I was broken because when I was implementing returns I
checked for functions after checking for operators always. This worked
for functions in operator use. However, with operators in functions it
would look for operators and find one inside the function call then
segfault when it tried to parse the operator with the closing
parentheses.