Parser: Moving Away From Using Tokenizer Everywhere
Previously I prefered using the tokenizer to handle the token based
functions (e.g. getting all the remaining tokens on a line). This made
the code look clean however it is very practical and results in not very
modular functions. I'm now moving toward passing a vector of tokens and
an index reference.