Parsed Token: Added variable_creation Parsed Token
This usurps the old variable creation method of parsing a type token and
just kinda assuming that the next thing is:
1. a variable name
2. only a variable name
We worked around both of these but it meant that in any case other than
parsing tokens as a statement. We could not create a variable. This is
very bad. Adding the variable_creation pt means that anywhere
expressions can be expressed a variable can be created. YAY!