Conditional Call: Find the Value of the Arguments Before Passing Them
This did not cause any issues that I noticed until the Variable Creation
changes I just made. In that case the for loop would not have any value
and we would segfault. This was very confusing especially because both
operator calls and function call DO find the value of the arguments
before executing the function/operator call.

We can now use the very clean (in my opinion) for syntax 'for (int index in start..end)'.
Before this change the variable declaration would have to be on a separate line
and at that point why not just use a while loop.